backend/requirements.txt
Django==4.2.11
djangorestframework==3.15.1
psycopg[binary]==3.1.18
django-cors-headers==4.3.1
